feat: optimize bundle size with tsup bundler
- Replace TypeScript compilation with tsup for optimized bundling
- Reduce bundle size from 20KB (3 files) to 16KB (single bundle)
- Add tsup.config.ts for esbuild-powered bundling configuration
- Update build scripts: npm run build (production) vs build:dev (debug)
- Fix test files to work with bundled structure
- External dependencies (@typescript-eslint/utils, eslint) for tree-shaking
- Update documentation with new build process and bundle details

Benefits:
- 20% smaller bundle size
- Single file distribution (lib/index.js + index.d.ts)
- Much faster build times (4ms vs ~500ms)
- Simplified distribution and installation
